translation_scripts/requirements.txt
openai>=1.0.0
tqdm>=4.64.0
python-dotenv>=1.0.0
asyncio
aiohttp
